+++ src/AdventOfCode/2018/Day_04.cs
@@ -5,17 +5,17 @@ public class Day_04
 {
     [Example(answer: 10 * 24, Example._1)]
     [Puzzle(answer: 26281, O.μs100)]
-    public long part_one(Lines lines) => Occurences(lines)
+    public int part_one(Lines lines) => Occurences(lines)
         .OrderByDescending(c => c.Value.Total).Select(Score).First();
 
     [Example(answer: 99 * 45, Example._1)]
     [Puzzle(answer: 73001, O.s)]
-    public long part_two(Lines lines) => Occurences(lines)
+    public int part_two(Lines lines) => Occurences(lines)
         .OrderByDescending(c => AsleepMostMinutes(c.Value)).Select(Score).First();
 
     static long AsleepMostMinutes(ItemCounter<int> o) => o.OrderByDescending(kvp => kvp.Count).First().Count;
     static int MinuteAsleepMost(ItemCounter<int> o) => o.OrderByDescending(kvp => kvp.Count).Select(kvp => kvp.Item).First();
-    static long Score(KeyValuePair<int, ItemCounter<int>> o) => o.Key* MinuteAsleepMost(o.Value);
+    static int Score(KeyValuePair<int, ItemCounter<int>> o) => o.Key* MinuteAsleepMost(o.Value);
 
     static Dictionary<int, ItemCounter<int>> Occurences(Lines lines)
     {
