+++ src/AdventOfCode/2021/Day_02.cs
@@ -5,15 +5,15 @@ public class Day_02
 {
     [Example(answer: 150, "forward 5;down 5;forward 8;up 3;down 8;forward 2")]
     [Puzzle(answer: 2027977, O.μs10)]
-    public long part_one(Lines lines) => new Submarine().Transpose(lines.As(Instruction.Parse), (s, i) => s.One(i)).Last().Produces;
+    public int part_one(Lines lines) => new Submarine().Transpose(lines.As(Instruction.Parse), (s, i) => s.One(i)).Last().Produces;
 
     [Example(answer: 900, "forward 5;down 5;forward 8;up 3;down 8;forward 2")]
     [Puzzle(answer: 1903644897, O.μs10)]
-    public long part_two(Lines lines) => new Submarine().Transpose(lines.As(Instruction.Parse), (s, i) => s.Two(i)).Last().Produces;
+    public int part_two(Lines lines) => new Submarine().Transpose(lines.As(Instruction.Parse), (s, i) => s.Two(i)).Last().Produces;
 
-    record Submarine(long Depth = 0, long Height = 0, long Aim = 0) 
+    record Submarine(int Depth = 0, int Height = 0, int Aim = 0) 
     {
-        public long Produces => Height * Depth;
+        public int Produces => Height * Depth;
         public Submarine One(Instruction i) => i.Type switch
         {
             'd' => this with { Depth = Depth + i.Velocity },
