+++ src/AdventOfCode/2022/Day_24.cs
@@ -6,7 +6,7 @@ public class Day_24
     [Example(answer: 18, Example._1)]
     [Example(answer: 255, Example._2)]
     [Puzzle(answer: 295, O.ms10)]
-    public long part_one(CharPixels chars)
+    public int part_one(CharPixels chars)
     {
         var state = State.Parse(chars);
         return state.Nav(state.P1, state.P2).Time;
@@ -15,7 +15,7 @@ public class Day_24
     [Example(answer: 54, Example._1)]
     [Example(answer: 809, Example._2)]
     [Puzzle(answer: 851, O.ms100)]
-    public long part_two(CharPixels chars)
+    public int part_two(CharPixels chars)
     {
         var state = State.Parse(chars);
         return state.Nav(state.P1, state.P2).Nav(state.P2, state.P1).Nav(state.P1, state.P2).Time;
